Split pane tree equalization logic
